backend

+++ lunchuis/src/main/java/com/project/lunchuis/Repository/BuyRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BuyRepository extends JpaRepository<Buy, Integer> {
+public interface BuyRepository extends JpaRepository<Buy, Long> {
 }
